+++ esperanto-tools-libs/CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [SW-19967] Try connecting to the MP server several times before failing
 - [SW-20002] Avoid running exit handlers on forked processes
 - [SW-20203] Remove duplicate profiling of kernel launches
+- [SW-20211] Improve maximum arguments size supported by allowing to clobber the exception buffer
 
 ## [0.13.1]
 ### Fixed